Write logic for a "random" computer choice
This is split into two functions.
The getRandomInt(), which itself is written in two possible ways
one as function declaration getRandomIntDec() and the other one as
function expression getRandomIntExp().
The returned value result is than used in the getComputerChoice() function.
It returns either Rock, Paper or Scissors as a String.